[MOBILE-7296] Add a basic stack unwinder
* add C++ terminate handlers (WIP)
* add C++ demangler (WIP)
* add/update logging macros
* refactor signal-handlers.c to signal-handlers.cpp
* tests are TBA

This is a work in progress, but provides the basic framework for unwinding
the backtrace from a crash (or passed signal). It relies on the
libunwind library that is shipped with the NDK, and supports Android 5
and higher. There is still more work to do to set up the unwinder for
each supported architecture.

At present, the stacktrace is simply logged, but will eventually be
written to a buffer for file serialization (probably as Json).

FIXME: Unwinding appears to work only when debugging, as the process is killed
too quickly to gather the backtrace.