[MOBILE-7298] Add simple symbolication tools
These scripts are mainly for reference, to demonstrate
how each symbolication task is performed using tools available
in the NDK. As such, NDK version 21.4 or higher must be installed
and configured by exporting the NDK root directory name as `ANDROID_NDK`
to the environment.

The collect and post tasks will integrated into the Gradle plugin,
and will be run executed during build time.